bookingsystem/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-webpack-dev-server==1.0.0
djangorestframework==3.15.2
idna==3.10
progressbar2==4.5.0
python-dotenv==1.0.1
python-utils==3.9.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
